[Ionut] [[BR]] -Decoupled "Scene" and "SceneManager" from "InputManager" [[BR]] --All input events are routed through the "Kernel" now [[BR]] --Renamed "InputManager" to "InputInterface" [[BR]] -Decoupled "Application" class from the rest of the engine except for the "Kernel" and "GUI" (for resize callbacks) [[BR]] -Added cached resolution values to classes so no queries to "Application" are needed anymore [[BR]] -Added a missing option from "config.xml" to enable or disable animations [[BR]] -Forgot to commit changed scene xml's after the romanian->english translation [[BR]] -All window dimension references are renamed to resolution and all variables related to it are now U16's as floats are irrelevant (what's a half pixel? for examle) [[BR]] -GUI resize callback is now called from "GFXDevice" instead of "GLWrapper" (will be moved to "Kernel" soon together with the "PostFX" resize callback) [[BR]]
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@118 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 572135482604208e7f4c29f03e2fab680d57e116 [formerly c32911fb41ad2a75d0b681be9cf04514092e9cb4]
Former-commit-id: 3af51d92cfc16a5bc98094e04734e06a74e9fcd0

+++ assets/shaders/GLSL/water.glsl
@@ -29,7 +29,7 @@ void main(void)
 
 -- Fragment
 
-uniform vec2 screenDimension;
+uniform ivec2 screenDimension;
 uniform float noise_tile;
 uniform float noise_factor;
 uniform float time;
